+++ packages-ui/roosterjs-react/lib/emoji/type/EmojiPaneStyles.ts
@@ -1,0 +1,25 @@
+import { IStyleSet } from '@fluentui/react/lib/Styling';
+/**
+ * @internal
+ * EmojiPane Style classes
+ */
+export interface EmojiPaneStyle {
+    quickPicker: IStyleSet;
+    tooltip: IStyleSet;
+    emojiTextInput: IStyleSet;
+    partialList: IStyleSet;
+    fullListContent: IStyleSet;
+    fullListBody: IStyleSet;
+    fullList: IStyleSet;
+    roosterEmojiPane: IStyleSet;
+    emoji: IStyleSet;
+    navBarButton: IStyleSet;
+    navBarTooltip: IStyleSet;
+    selected: IStyleSet;
+    navBar: IStyleSet;
+    statusBarDetails: IStyleSet;
+    statusBarDetailsContainer: IStyleSet;
+    statusBar: IStyleSet;
+    statusBarIcon: IStyleSet;
+    statusBarNoResultDetailsContainer: IStyleSet;
+}
